requirements.txt
Django==4.1.6
djangorestframework==3.13.1
drf-spectacular==0.23.1
django-filter==22.1